requirements.txt
gunicorn
django >= 3.1.3
django-heroku
pillow
django-allauth